fix(#81): checking if the user is admin before allow to remove a user

+++ app/controllers/groupMembershipController.js
@@ -58,6 +58,7 @@ const getGroupMembers = async (req, res) => {
 
 const deleteGroupMembership = async (req, res) => {
     const { groupId, userId } = req.params;
+    const currentUserId = req.userId;
 
     try {
         const group = await Group.findById(groupId);
@@ -70,6 +71,13 @@ const deleteGroupMembership = async (req, res) => {
 
         if (!membership) {
             return res.status(404).json({ message: "L'utilisateur n'est pas membre du groupe" });
+        }
+
+        // Check if the current user is an administrator of the group
+        const isAdmin = await GroupMembership.exists({ user_id: currentUserId, group_id: groupId, is_administrator: true });
+
+        if (!isAdmin) {
+            return res.status(403).json({ message: "Vous n'êtes pas autorisé à supprimer des membres du groupe" });
         }
 
         await GroupMembership.findByIdAndDelete(membership._id);
